api/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
gunicorn==19.9.0pytz==2018.9
